Update GCodeGenerator.java
Modified the code so that it remembers where the last click was. Also created a method to do swipe gestures and a method to have the maker bot finish and go back to home. Some values (particularly the offsets) may need to be adjusted for accuracy depending on when the stylus is attached.